lnd+cmd/chantools: extract functions from walletinfo
